+++ molgenis-data/src/main/java/org/molgenis/data/support/exceptions/TemplateExpressionInvalidTagException.java
@@ -1,16 +1,16 @@
-package org.molgenis.data.support;
+package org.molgenis.data.support.exceptions;
 
 import static java.util.Objects.requireNonNull;
 
 import org.molgenis.util.exception.CodedRuntimeException;
 
-class TemplateExpressionInvalidTagException extends CodedRuntimeException {
+public class TemplateExpressionInvalidTagException extends CodedRuntimeException {
   private static final String ERROR_CODE = "D12c";
 
   private final String expression;
   private final String tag;
 
-  TemplateExpressionInvalidTagException(String expression, String tag) {
+  public TemplateExpressionInvalidTagException(String expression, String tag) {
     super(ERROR_CODE);
     this.expression = requireNonNull(expression);
     this.tag = requireNonNull(tag);
