+++ molgenis-data/src/main/java/org/molgenis/data/support/exceptions/TemplateExpressionMathInvalidParameterException.java
@@ -1,0 +1,9 @@
+package org.molgenis.data.support.exceptions;
+
+public class TemplateExpressionMathInvalidParameterException extends RuntimeException {
+  private static final String ERROR_CODE = "D12h";
+
+  public TemplateExpressionMathInvalidParameterException() {
+    super(ERROR_CODE);
+  }
+}
