+++ molgenis-data/src/test/java/org/molgenis/data/support/TemplateExpressionEvaluatorTest.java
@@ -26,6 +26,11 @@ import org.mockito.Mock;
 import org.molgenis.data.Entity;
 import org.molgenis.data.meta.model.Attribute;
 import org.molgenis.data.meta.model.EntityType;
+import org.molgenis.data.support.exceptions.TemplateExpressionException;
+import org.molgenis.data.support.exceptions.TemplateExpressionInvalidTagException;
+import org.molgenis.data.support.exceptions.TemplateExpressionMissingTagException;
+import org.molgenis.data.support.exceptions.TemplateExpressionSyntaxException;
+import org.molgenis.data.support.exceptions.TemplateExpressionUnknownAttributeException;
 import org.molgenis.test.AbstractMockitoTest;
 
 class TemplateExpressionEvaluatorTest extends AbstractMockitoTest {
